Major refactoring of types
