+++ src/types/account_info_type.py
@@ -1,0 +1,12 @@
+from graphene import List, ObjectType, String
+
+class TransactionType(ObjectType):
+  name = String(required=True)
+  timestamp = String(required=True)
+
+class AccountInfoType(ObjectType):
+  brbs = String(required=True)
+  city_bucks = String(required=True)
+  history = List(TransactionType, required=True)
+  laundry = String(required=True)
+  swipes = String(required=True)
